backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dases"
version = "0.1.0"
description = "基于计算机科学领域学术论文的多智能体检索系统"
requires-python = ">=3.11"
dependencies = [
    # --- Web 框架 ---
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.0",
    # --- 数据库 ---
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # --- Redis ---
    "redis>=5.0",
    # --- LLM / Agent ---
    "langchain>=0.3",
    "langgraph>=0.2",
    "litellm>=1.50",
    # --- 向量数据库 ---
    "pymilvus>=2.4",
    # "qdrant-client>=1.12",       # 备选
    # --- Embedding ---
    "sentence-transformers>=3.0",
    "torch>=2.0",  # CPU-only, source overridden via [tool.uv.sources]
    # --- 文档解析 ---
    "pymupdf>=1.24",
    "unstructured>=0.15",
    # --- 可观测性 ---
    "langfuse>=2.50",
    # --- Elasticsearch ---
    "elasticsearch[async]>=8.13,<9",
    # --- 工具 ---
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rank-bm25>=0.2",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
    "mypy>=1.11",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]



[tool.uv]
torch-backend = "cpu"

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
